+++ CHANGELOG.md
@@ -14,6 +14,14 @@ Types of changes
 - `Fixed` for any bug fixes.
 - `Security` in case of vulnerabilities.
 
+## [2.0.0] (Ureleased)
+
+- `Added` configuration of exported columns in tables.yaml, see issue #33 for more information.
+- `Added` MariaDB/MySQL support (thanks to @joaking85).
+- `Added` auto-select columns required by a relation but not exported in tables.yaml.
+- `Changed` order of keys in output JSON lines will be alphabetical when pulling (without configuration in tables.yaml).
+- `Fixed` limit keyword on DB2 dialect.
+
 ## [1.9.0]
 
 - `Added` new verb to extract, to get status and to update sequences
@@ -37,7 +45,7 @@ Types of changes
 
 ## [1.4.0]
 
-- `Added` statistics report for push and pull executions
+- `Added` statistics report for push and pull executions (thanks to @CapKicklee)
 - `Changed` some info level logs to debug level in pull module
 
 ## [1.3.1]
@@ -55,7 +63,7 @@ Types of changes
 
 ## [1.2.0]
 
-- `Added` structured logs (debug & json format)
+- `Added` structured logs (debug & json format) (thanks to @CapKicklee)
 
 ## [1.1.2]